Use ms instead of secs for all calculations
Using seconds for the calculations causes apply_tracking() to set the normal speed at the start of each second but as time passes stepsLeft decreases but timeLeft remains the same. As a result the motor slows down. When the start of the new second is reached, the speed bounces up to normal again and the pattern repeats. The effect is stronger near the end of the 15 second planning time.