Implement new position calculation method
The new calculation method allows the two legs of the isosceles triangle to be of different length (i.e. t is not really isosceles)
This will reduce the error when the construction of the barn door is not as accurate.
Also switched to using start and end opening distance instead of angles as distances are easier to measure